refactor: swagger
