+++ package.json
@@ -7,14 +7,17 @@
   "author": "Lipoic backend team",
   "license": "GPL-3.0",
   "scripts": {
-    "build": "yarn swagger && tsc && tsc-alias",
-    "dev": "yarn swagger && concurrently \"tsc -w\" \"tsc-alias -w\" \"nodemon dist/src/index.js\"",
+    "build": "tsc && tsc-alias",
+    "dev:build": "tsc -w",
+    "dev:build-alias": "tsc-alias -w",
+    "dev:run": "nodemon dist/src/index.js",
+    "dev": "concurrently yarn:dev:*",
     "start": "yarn build && node dist/src/index.js",
     "lint": "eslint src/**/*.ts --fix",
     "test": "yarn swagger && vitest --no-threads",
     "test:ui": "yarn test --ui",
     "coverage": "yarn swagger && vitest run --no-threads --coverage",
-    "swagger": "ts-node swagger.ts"
+    "swagger": "tsc && tsc-alias src/util/swagger.ts && node dist/src/util/swagger.js"
   },
   "dependencies": {
     "app-root-path": "^3.1.0",
@@ -58,5 +61,13 @@
     "tsc-alias": "1.7.0",
     "typescript": "4.8.3",
     "vitest": "0.23.2"
+  },
+  "nodemonConfig": {
+    "ext": "js",
+    "ignore": [
+      "*.test.ts",
+      "swagger-output.json"
+    ],
+    "delay": "2"
   }
 }
