+++ tsconfig.json
@@ -4,9 +4,15 @@
     "module": "commonjs",
     "baseUrl": ".",
     "paths": {
-      "@/*": ["src/*"],
-      "#/*": ["src/router/*"],
-      "#": ["src/router"]
+      "@/*": [
+        "src/*"
+      ],
+      "#/*": [
+        "src/router/*"
+      ],
+      "#": [
+        "src/router"
+      ]
     },
     "outDir": "./dist",
     "esModuleInterop": true,
@@ -15,6 +21,11 @@
     "strict": true,
     "skipLibCheck": true
   },
-  "include": ["src/**/*.ts", "test/**/*.ts", "swagger.ts"],
-  "exclude": ["node_modules"]
+  "include": [
+    "src/**/*.ts",
+    "test/**/*.ts"
+  ],
+  "exclude": [
+    "node_modules"
+  ]
 }
