+++ vitest.config.ts
@@ -19,8 +19,8 @@ export default defineConfig({
         'src/index.ts',
         '.eslintrc.js',
         'vitest.config.ts',
-        'swagger.ts',
         '**/*.d.ts',
+        'src/util/swagger.ts',
       ],
     },
   },
